workflow: 1) add the first work flow to verify the push and pull merge code 2) correct some error in the README 3) rename the quantum module to gaussian, move it to the dir hotpot.qm
