+++ doc/API/cheminfo.md
@@ -1,0 +1,14 @@
+# hotpot.cheminfo module
+this module to handle Chemoinformatics
+
+## Crystal
+
+## Molecule
+
+## Atom
+
+## Bond
+
+## Angle
+
+## Torsion
