+++ doc/API/qm.gaussian.md
@@ -1,0 +1,3 @@
+# hotpot.qm.gaussian module
+
+![Code structure of qm.gaussian](../pictures/Gauss.png)
